Offset changes for attachment points
Changed the hoveroffset initial values so that instead of just having a y-offset the attachment points would move around the top or bottom of the atom. Must be revised on atoms with more attachment points because we might want to spread them up and down equally but I don't know how to test it with only two input attachments. The hovered attachments move up and to the right unless it is too close to the top and then they switch to dropping down and to the right.