Update prototypes.js
Change offset so that it moves around the parent molecule.
Angle and offset numbers should probably be revised since they were just tests for the translate molecule input menu.
Next: First attachment point should probably change if theres 2 attachments or 6, right now it looks best with for but maybe it being responsive to the number of IOs would be nice